Feature/delete post

+++ server/controllers/post.js
@@ -56,43 +56,45 @@ exports.update = function *() {
       },
       include: [ { model: models.Tag } ]
     });
+    if( UserId === post.CreaterId || UserId === post.EditorId )
+    {
+      // Remove Tag
 
-    // Remove Tag
+      // yield * foreach(post.Tags, function * (tag, index) {
 
-    // yield * foreach(post.Tags, function * (tag, index) {
-
-    yield post.Tags.map((tag) => {
-      let state = editPost.tags.indexOf(tag.name);
-      if(state === -1){
-        // New Post Data not have this tag, Remove Tag.
-        models.Tag.destroy({
-          where:{
-            id:tag.id
-          }
-        });
-      }else {
-        // Is exist remove in editTag
-        editPost.tags.splice(state,1);
-      }
-    });
-
-    yield editPost.tags.map((tag) => {
-      // Create new Tag
-      models.Tag.create({
-        name:tag,
-        PostId:post.id
+      yield post.Tags.map((tag) => {
+        let state = editPost.tags.indexOf(tag.name);
+        if(state === -1){
+          // New Post Data not have this tag, Remove Tag.
+          models.Tag.destroy({
+            where:{
+              id:tag.id
+            }
+          });
+        }else {
+          // Is exist remove in editTag
+          editPost.tags.splice(state,1);
+        }
       });
-    });
 
-    // Post
-    post.title=editPost.title;
-    post.content=editPost.content;
-    post.img=editPost.img;
+      yield editPost.tags.map((tag) => {
+        // Create new Tag
+        models.Tag.create({
+          name:tag,
+          PostId:post.id
+        });
+      });
 
-    post.UserId = UserId;
-    result = yield post.save();
+      // Post
+      post.title=editPost.title;
+      post.content=editPost.content;
+      post.img=editPost.img;
 
-    this.body = {result};
+      result = yield post.save();
+
+      this.body = {result};
+    }
+
   } catch (error) {
     console.log(error.stack);
     this.body = {result, error};
@@ -101,6 +103,32 @@ exports.update = function *() {
 
 };
 
+exports.updateEditor = function *() {
+
+
+  try {
+    let authStatus = services.user.getAuthStatus(this);
+    if(authStatus.authority ==='admin'){
+      let postId = this.params.id;
+      let editorId = this.request.body.editorId;
+      let result = null;
+
+      if(editorId==='0'){
+        editorId=null;
+      };
+
+      let post = yield models.Post.findById(postId);
+      post.EditorId = editorId;
+      yield post.save();
+
+      this.body = {post};
+    }
+  } catch (error) {
+    console.log(error.stack);
+    this.body = {result, error};
+  }
+
+};
 
 exports.upload = function* (next) {
 
@@ -136,13 +164,19 @@ exports.upload = function* (next) {
 
 exports.delete = function *() {
 
+
+
   let postId = this.params.id;
 
   let result = null;
 
   try {
+    let UserId = services.user.getSessionUser(this).id;
     let post = yield models.Post.findById(postId);
-    result = post.destroy()
+
+    if( UserId === post.CreaterId || UserId === post.EditorId )
+      result = yield post.destroy()
+
   } catch (e) {
     console.error("delete post error", e);
   }
